Section 6: Lecture 134: Learnt Layers
Learnt How to use Layers in Blender

First, in User Preferences --> Add-On --> Turn On Layer Management

Layers, In Blender are represent the (1 to 0) from the keyboard top of the number key not from numpad
In Blender, you can see the small 20 box, that is Layers

In Object Mode --> Object -- Move (m), you can move the object to specify layer.
or
Select the Object and click the layer dot to move it.

and more, please review if forgot