Section 5, Lecture 139: Particles System, Make Hair
Learnt what is Particles and how to use.

In Properties menu --> Particles Menu --> Add

Type can be change to Hair (it make hair)
Children are spawn around the parent
Children are base on the Total Number of Emission multipler it
From eg, total is 1000. And Children Number is 10.
1000 * 10 = 10000;

Interpolated are better than Simple in Children type.

Create Body Hair for Rabbit